Use `int` for integer values instead of `short`

+++  NumeralConverter/ NumeralConverter/Program.cs
@@ -8,7 +8,7 @@ while (true)
     if (string.Equals(input, "q", StringComparison.OrdinalIgnoreCase))
         break;
 
-    Console.WriteLine(short.TryParse(input, out var number)
+    Console.WriteLine(int.TryParse(input, out var number)
         ? $"Roman numeral: {Converter.ConvertToRoman(number)}"
         : "Invalid input, please enter a number between 1 and 3999.");
 }
